requirements.txt
# Updated requirements.txt - Windows compatible
streamlit
langchain
PyPDF2
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
pandas
numpy